Add:댓글 삭제 기능 구현(전체허용상태)

+++ src/main/java/com/mbc/sharetravel_spring/posts/Comment.java
@@ -28,7 +28,7 @@ public class Comment {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // 댓글 고유번호
+    private Integer id; // 댓글 고유번호
 
     @Column(nullable = false)
     private String content; // 댓글 내용
